src/plain_python/requirements.txt
confluent-kafka==2.1.1
kafka-python==2.0.2
